requirements.txt
apache-airflow==2.8.3
pandas>=1.5.0
scikit-learn>=1.2.0
joblib>=1.2.0
requests>=2.28.0
